Refactoring + reorder/rename folders & files
