+++ src/components/landing/Landing.tsx
@@ -1,31 +1,29 @@
 import Filters from '../filters/Filters';
-import MenuNavbar from '../sideNavbar/MenuNavbar';
+import MenuNavbar from '../menuNavbar/MenuNavbar';
 import Students from '../students/Students';
-import UserNavbar from './UserNavbar';
+import UserNavbar from '../userNavBar/UserNavbar';
 import UserDetails from '../userDetailsPanel/UserDetails';
 
 const Landing = () => {
   return (
     <div className="flex">
       <MenuNavbar />
-      <div className="flex-1">
+      <div className="flex w-full flex-col gap-3 p-2 pb-8 pr-14">
         {/* login nav to the up-right */}
         <div className="flex justify-end">
           <UserNavbar />
         </div>
 
-        <main className="flex">
+        <div className="flex">
           {/* White board --> */}
-          <div className="min-h-[90vh] w-full rounded-[15px] bg-white md:mr-8">
-            <div className="relative flex gap-1 pb-2 pl-12 pt-12">
-              <Filters />
-              <Students />
-            </div>
+          <div className="flex w-full min-h-screen gap-2 rounded-xl bg-white p-10">
+            <Filters />
+            <Students />
           </div>
 
           {/* Profile Details */}
           <UserDetails />
-        </main>
+        </div>
       </div>
     </div>
   );
