+++ src/components/userNavBar/UserNavbar.tsx
@@ -37,13 +37,13 @@ const UserNavbar = () => {
   };
 
   return (
-    <div className="mx-4 my-3 flex w-full justify-between md:mr-8 md:justify-end">
+    <div className="flex w-full justify-between md:justify-end">
       <div className="flex cursor-pointer items-center justify-center md:hidden ">
         <img src={burgerMenu} alt="burger menu" className="h-8 w-8 " />
       </div>
 
-      <div className="dropdown dropdown-end mx-4 rounded-full bg-white font-medium hover:bg-gray-300 md:rounded-md">
-        <div tabIndex={0} role="button" className="w-full px-3 py-2">
+      <div className="hover:bg-gray-300 dropdown dropdown-end mx-4 rounded-full bg-white font-medium md:rounded-md">
+        <div tabIndex={0} role="button" className="w-full p-2">
           <div className="mx-0.1 relative flex items-center">
             <span className="text-gray">ES</span>
             <img src={ArrowDown} alt="Vector" className="ml-1 mt-0.5 h-2 w-4" />
@@ -64,10 +64,10 @@ const UserNavbar = () => {
       </div>
 
       <button
-        className="text-gray flex items-center justify-center rounded-full bg-white font-medium hover:bg-gray-300 md:rounded-md"
+        className="hover:bg-gray-300 flex items-center justify-center rounded-full bg-white font-medium text-gray md:rounded-md"
         onClick={handleOpenRestrictedPopup}
       >
-        <p className="text-gray w-full px-3 py-2">Login/registro</p>
+        <p className="w-full px-3 py-2 text-gray">Login/registro</p>
       </button>
       {isRestrictedPopupOpen && (
         <div className="fixed left-0 top-0 z-50 flex h-full w-full items-center justify-center bg-black bg-opacity-50">
